Can select 1st category
This patch adds a button to validate the form. Without it
the form is auto validated "on change". But it was hence not
possible to select the 1st category from this page.
With this patch is now possible to easily select this category.

Since there is now a button, it becomes counter-intuitive if the
form can be auto-validated, so this patch remove the submit
on-change