admin can edit plugin_tag_desc_sidebar_html
The admin form only allows editions of the options plugin_cat_desc_font_size
and plugin_car_desc_permit_edit. It's not clear to me how an admin is supposed to
edit the value of the option plugin_tag_desc_sidebar_html.
This commit makes it possible to edit this option from the administration dashboard